update feature product, entity promotion and user_promotion

+++ backend/src/main/java/com/logistics/dto/PromotionDto.java
@@ -18,7 +18,7 @@ public class PromotionDto {
     private String description;
     private String discountType;
     private BigDecimal discountValue;
-    private Integer minOrderValue;
+    private BigDecimal minOrderValue;
     private Integer maxDiscountAmount;
     private LocalDateTime startDate;
     private LocalDateTime endDate;
